+++ src/config.rs
@@ -144,6 +144,8 @@ pub struct Workers {
 pub struct Middlewares {
     /// Middleware that enable compression for the response.
     pub compression: Option<EnableMiddleware>,
+    /// Middleware that enable etag cache headers.
+    pub etag: Option<EnableMiddleware>,
     /// Middleware that limit the payload request.
     pub limit_payload: Option<LimitPayloadMiddleware>,
     /// Middleware that improve the tracing logger and adding trace id for each